pyproject.toml
[project]
name = "diagnostico-de-rede"
version = "0.1.0"
description = ""
authors = [
    {name = "FelipeCN",email = "felipe@felipecncloud.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "speedtest-cli (>=2.1.3,<3.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "fpdf (>=1.7.2,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "requests (>=2.32.4,<3.0.0)",
    "nbt (>=1.5.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
